# Allowed inlining for traverseWithIndex