@treeowl 2019-04-11
`INLINABLE` should be subject to phase controls. Do they not work out the way we need them to?