@oisdk 2019-04-11
I hadn't realised that! re-running the benchmarks now with `INLINABLE`