@oisdk 2019-04-11
Nope! I'll take a look now. Should I put an `#ifdef __GLASGOW_HASKELL__` around the `INLINABLE`?